Moved external texture to Plugins, renamed GraphicsInterops to ExternalTexture
