+++ Apps/Playground/CMakeLists.txt
@@ -118,7 +118,6 @@ target_link_to_dependencies(Playground
     PRIVATE ScriptLoader
     PRIVATE XMLHttpRequest
     PRIVATE Canvas
-    PRIVATE GraphicsInterops
     ${ADDITIONAL_LIBRARIES}
     ${BABYLON_NATIVE_PLAYGROUND_EXTENSION_LIBRARIES})
