+++ Core/CMakeLists.txt
@@ -2,4 +2,3 @@ add_subdirectory(JsRuntime)
 add_subdirectory(AppRuntime)
 add_subdirectory(ScriptLoader)
 add_subdirectory(Graphics)
-add_subdirectory(GraphicsInterops)
